Adding Akka-cluster sample example for protocol-buffer serialization
