--- akka-sample-cluster-scala-protobuf/LICENSE
@@ -1,10 +1,0 @@
-Akka sample by Lightbend
-
-Licensed under Public Domain (CC0)
-
-To the extent possible under law, the person who associated CC0 with
-this Template has waived all copyright and related or neighboring
-rights to this Template.
-
-You should have received a copy of the CC0 legalcode along with this
-work.  If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
