--- akka-sample-cluster-scala-protobuf/build.sbt
@@ -1,36 +1,0 @@
-import com.typesafe.sbt.SbtMultiJvm.multiJvmSettings
-import com.typesafe.sbt.SbtMultiJvm.MultiJvmKeys.MultiJvm
-
-val akkaVersion = "2.6.0-M1"
-
-lazy val `akka-sample-cluster-scala` = project
-  .in(file("."))
-  .settings(multiJvmSettings: _*)
-  .settings(
-    organization := "com.typesafe.akka.samples",
-    scalaVersion := "2.12.8",
-    scalacOptions in Compile ++= Seq("-deprecation", "-feature", "-unchecked", "-Xlog-reflective-calls", "-Xlint"),
-    javacOptions in Compile ++= Seq("-Xlint:unchecked", "-Xlint:deprecation"),
-    javaOptions in run ++= Seq("-Xms128m", "-Xmx1024m", s"-Djava.library.path=${baseDirectory}/target/native"),
-    libraryDependencies ++= Seq(
-      "com.typesafe.akka" %% "akka-actor" % akkaVersion,
-      "com.typesafe.akka" %% "akka-remote" % akkaVersion,
-      "com.typesafe.akka" %% "akka-cluster" % akkaVersion,
-      "com.typesafe.akka" %% "akka-cluster-metrics" % akkaVersion,
-      "com.typesafe.akka" %% "akka-cluster-tools" % akkaVersion,
-      "com.typesafe.akka" %% "akka-multi-node-testkit" % akkaVersion,
-      "com.thesamet.scalapb" %% "scalapb-runtime" % scalapb.compiler.Version.scalapbVersion % "protobuf",
-      "org.scalatest" %% "scalatest" % "3.0.7" % Test,
-      "io.kamon" % "sigar-loader" % "1.6.5-rev003"),
-    fork in run := true,
-    mainClass in (Compile, run) := Some("sample.cluster.transformation.TransformationApp"),
-    // disable parallel tests
-    parallelExecution in Test := false,
-    licenses := Seq(("CC0", url("http://creativecommons.org/publicdomain/zero/1.0")))
-  )
-  .configs (MultiJvm)
-
-
-PB.targets in Compile := Seq(
-  scalapb.gen() -> (sourceManaged in Compile).value
-)
