--- akka-sample-cluster-scala-protobuf/project/plugins.sbt
@@ -1,2 +1,0 @@
-addSbtPlugin("com.typesafe.sbt" % "sbt-multi-jvm" % "0.4.0")
-addSbtPlugin("com.dwijnand" % "sbt-dynver" % "3.0.0")
