--- akka-sample-cluster-scala-protobuf/project/protoc.sbt
@@ -1,2 +1,0 @@
-addSbtPlugin("com.thesamet" % "sbt-protoc" % "0.99.23")
-libraryDependencies += "com.thesamet.scalapb" %% "compilerplugin" % "0.9.0"
