--- akka-sample-cluster-scala-protobuf/src/main/resources/application.conf
@@ -1,48 +1,0 @@
-akka {
-  actor {
-    provider = cluster
-    allow-java-serialization = off
-    serializers {
-      scalapb = "sample.cluster.transformation.ScalaPbSerializer"
-    }
-    serialization-bindings {
-      "scalapb.GeneratedMessage" = scalapb
-    }
-    serialization-identifiers {
-      "sample.cluster.transformation.ScalaPbSerializer" = 10000
-    }
-  }
-
-  remote {
-    classic.netty.tcp {
-      hostname = "127.0.0.1"
-      port = 0
-    }
-
-    artery {
-      # change this to enabled=on to use Artery instead of netty
-      # see https://doc.akka.io/docs/akka/current/remoting-artery.html
-      enabled = off
-      transport = tcp
-      canonical.hostname = "127.0.0.1"
-      canonical.port = 0
-    }
-  }
-
-  cluster {
-    seed-nodes = [
-      "akka.tcp://ClusterSystem@127.0.0.1:2551",
-      "akka.tcp://ClusterSystem@127.0.0.1:2552"]
-
-    # auto downing is NOT safe for production deployments.
-    # you may want to use it during development, read more about it in the docs.
-    auto-down-unreachable-after = 10s
-  }
-}
-
-# Enable metrics extension in akka-cluster-metrics.
-akka.extensions=["akka.cluster.metrics.ClusterMetricsExtension"]
-
-# Sigar native library extract location during tests.
-# Note: use per-jvm-instance folder when running multiple jvm on one host.
-akka.cluster.metrics.native-library-extract-folder=${user.dir}/target/native
