--- akka-sample-cluster-scala-protobuf/src/main/scala/sample/cluster/transformation/TransformationBackend.scala
@@ -1,54 +1,0 @@
-package sample.cluster.transformation
-
-import akka.actor.Actor
-import akka.actor.ActorSystem
-import akka.actor.Props
-import akka.actor.RootActorPath
-import akka.cluster.Cluster
-import akka.cluster.ClusterEvent.CurrentClusterState
-import akka.cluster.ClusterEvent.MemberUp
-import akka.cluster.Member
-import akka.cluster.MemberStatus
-import com.typesafe.config.ConfigFactory
-import sample.cluster.transformation.TransformationMessages._
-
-//#backend
-class TransformationBackend extends Actor {
-
-  val cluster = Cluster(context.system)
-
-  // subscribe to cluster changes, MemberUp
-  // re-subscribe when restart
-  override def preStart(): Unit = cluster.subscribe(self, classOf[MemberUp])
-  override def postStop(): Unit = cluster.unsubscribe(self)
-
-  def receive: PartialFunction[Any, Unit] = {
-    case TransformationJob(text) => sender() ! TransformationResult(text.toUpperCase)
-    case state: CurrentClusterState =>
-      state.members.filter(_.status == MemberStatus.Up) foreach register
-    case MemberUp(m) => register(m)
-  }
-
-  def register(member: Member): Unit =
-    if (member.hasRole("frontend"))
-      context.actorSelection(RootActorPath(member.address) / "user" / "frontend") !
-        BackendRegistration()
-}
-//#backend
-
-object TransformationBackend {
-  def main(args: Array[String]): Unit = {
-    // Override the configuration of the port when specified as program argument
-    // To use artery instead of netty, change to "akka.remote.artery.canonical.port"
-    // See https://doc.akka.io/docs/akka/current/remoting-artery.html for details
-    val port = if (args.isEmpty) "0" else args(0)
-    val config = ConfigFactory.parseString(s"""
-        akka.classic.remote.netty.tcp.port=$port
-        """)
-      .withFallback(ConfigFactory.parseString("akka.cluster.roles = [backend]"))
-      .withFallback(ConfigFactory.load())
-
-    val system = ActorSystem("ClusterSystem", config)
-    system.actorOf(Props[TransformationBackend], name = "backend")
-  }
-}
