--- akka-sample-cluster-scala-protobuf/src/main/scala/sample/cluster/transformation/TransformationFrontend.scala
@@ -1,65 +1,0 @@
-package sample.cluster.transformation
-
-import language.postfixOps
-import scala.concurrent.duration._
-import akka.actor.Actor
-import akka.actor.ActorRef
-import akka.actor.ActorSystem
-import akka.actor.Props
-import akka.actor.Terminated
-import akka.pattern.ask
-import akka.util.Timeout
-import com.typesafe.config.ConfigFactory
-import java.util.concurrent.atomic.AtomicInteger
-
-import sample.cluster.transformation.TransformationMessages._
-
-//#frontend
-class TransformationFrontend extends Actor {
-
-  var backends = IndexedSeq.empty[ActorRef]
-  var jobCounter = 0
-
-  def receive = {
-    case job: TransformationJob if backends.isEmpty =>
-      sender() ! JobFailed("Service unavailable, try again later", Some(job))
-
-    case job: TransformationJob =>
-      jobCounter += 1
-      backends(jobCounter % backends.size) forward job
-
-    case _: BackendRegistration if !backends.contains(sender()) =>
-      context watch sender()
-      backends = backends :+ sender()
-
-    case Terminated(a) =>
-      backends = backends.filterNot(_ == a)
-  }
-}
-//#frontend
-
-object TransformationFrontend {
-  def main(args: Array[String]): Unit = {
-    // Override the configuration of the port when specified as program argument
-    // To use artery instead of netty, change to "akka.remote.artery.canonical.port"
-    // See https://doc.akka.io/docs/akka/current/remoting-artery.html for details
-    val port = if (args.isEmpty) "0" else args(0)
-    val config = ConfigFactory.parseString(s"""
-        akka.remote.classic.netty.tcp.port=$port
-        """)
-      .withFallback(ConfigFactory.parseString("akka.cluster.roles = [frontend]"))
-      .withFallback(ConfigFactory.load())
-    val system = ActorSystem("ClusterSystem", config)
-    val frontend = system.actorOf(Props[TransformationFrontend], name = "frontend")
-
-    val counter = new AtomicInteger
-    import system.dispatcher
-    system.scheduler.schedule(2.seconds, 2.seconds) {
-      implicit val timeout = Timeout(5 seconds)
-      (frontend ? TransformationJob("hello-" + counter.incrementAndGet())) foreach  {
-        case result => println(result)
-      }
-    }
-
-  }
-}
