--- akka-sample-cluster-scala-protobuf/src/test/resources/reference.conf
@@ -1,4 +1,0 @@
-# Don't terminate ActorSystem in tests
-akka.coordinated-shutdown.run-by-jvm-shutdown-hook = off
-akka.coordinated-shutdown.terminate-actor-system = off
-akka.cluster.run-coordinated-shutdown-when-down = off
