+++ akka-sample-cluster-scala/README.md
@@ -59,7 +59,29 @@ Let's take a look at an example that illustrates how workers, here named *backen
 
 The example application provides a service to transform text. When some text is sent to one of the frontend services, it will be delegated to one of the backend workers, which performs the transformation job, and sends the result back to the original client. New backend nodes, as well as new frontend nodes, can be added or removed to the cluster dynamically.
 
-Open [TransformationMessages.scala](src/main/scala/sample/cluster/transformation/TransformationMessages.scala). It defines the messages that are sent between the actors.
+Open [TransformationMessages.proto](src/main/protobuf/sample/cluster/transformation/TransformationMessages.proto). It defines the messages that are sent between the actors. And we use scala-pb to generate the class files for protocol-buffer based classes for actor messages.
+
+The [Scalapb](https://github.com/scalapb/ScalaPB), automatically generates class files for *.proto files present inside src/main/protobuf folder.
+
+And the messages are serialized using [ScalaPbSerializer](src/main/scala/sample/cluster/transformation/ScalaPbSerializer.scala), which need to be specified as part of the application.conf as shown below.
+
+```
+   akka {
+      actor {
+        provider = cluster
+        allow-java-serialization = off
+        serializers {
+          scalapb = "sample.cluster.transformation.ScalaPbSerializer"
+        }
+        serialization-bindings {
+          "scalapb.GeneratedMessage" = scalapb
+        }
+        serialization-identifiers {
+          "sample.cluster.transformation.ScalaPbSerializer" = 10000
+        }
+      }
+  }
+```
 
 The backend worker that performs the transformation job is defined in [TransformationBackend.scala](src/main/scala/sample/cluster/transformation/TransformationBackend.scala).
