+++ akka-sample-cluster-scala/project/plugins.sbt
@@ -1,2 +1,5 @@
 addSbtPlugin("com.typesafe.sbt" % "sbt-multi-jvm" % "0.4.0")
 addSbtPlugin("com.dwijnand" % "sbt-dynver" % "3.0.0")
+
+addSbtPlugin("com.thesamet" % "sbt-protoc" % "0.99.23")
+libraryDependencies += "com.thesamet.scalapb" %% "compilerplugin" % "0.9.0"
