+++ akka-sample-cluster-scala/src/main/protobuf/TransformationMessages.proto
