+++ akka-sample-cluster-scala/src/main/resources/application.conf
@@ -1,9 +1,20 @@
 akka {
   actor {
     provider = cluster
+    allow-java-serialization = off
+    serializers {
+      scalapb = "sample.cluster.transformation.ScalaPbSerializer"
+    }
+    serialization-bindings {
+      "scalapb.GeneratedMessage" = scalapb
+    }
+    serialization-identifiers {
+      "sample.cluster.transformation.ScalaPbSerializer" = 10000
+    }
   }
+
   remote {
-    netty.tcp {
+    classic.netty.tcp {
       hostname = "127.0.0.1"
       port = 0
     }
