+++ akka-sample-cluster-scala/src/main/scala/sample/cluster/transformation/ScalaPbSerializer.scala
