--- akka-sample-cluster-scala/src/main/scala/sample/cluster/transformation/TransformationMessages.scala
@@ -1,8 +1,0 @@
-package sample.cluster.transformation
-
-//#messages
-final case class TransformationJob(text: String)
-final case class TransformationResult(text: String)
-final case class JobFailed(reason: String, job: TransformationJob)
-case object BackendRegistration
-//#messages
